Refactor modules and enforce naming conventions
