+++ pogo_analyzer/__init__.py
@@ -1,15 +1,29 @@
 """Utilities shared across the raid scoreboard tooling."""
 
+from __future__ import annotations
+
+from .data.raid_entries import (
+    DEFAULT_RAID_ENTRIES,
+    IVSpread,
+    PokemonRaidEntry,
+    build_entry_rows,
+)
+from .raid_entries import RAID_ENTRIES, build_rows
+from .scoring import calculate_iv_bonus, calculate_raid_score, iv_bonus, raid_score
 from .simple_table import Row, SimpleSeries, SimpleTable
-from .raid_entries import RAID_ENTRIES, PokemonRaidEntry
-from .scoring import iv_bonus, raid_score
 
 __all__ = [
-    "RAID_ENTRIES",
+    "DEFAULT_RAID_ENTRIES",
+    "IVSpread",
     "PokemonRaidEntry",
+    "RAID_ENTRIES",
     "Row",
     "SimpleSeries",
     "SimpleTable",
+    "build_entry_rows",
+    "build_rows",
+    "calculate_iv_bonus",
+    "calculate_raid_score",
     "iv_bonus",
     "raid_score",
 ]
