+++ pogo_analyzer/scoring/__init__.py
@@ -1,0 +1,17 @@
+"""Public scoring helpers with consistent naming."""
+
+from __future__ import annotations
+
+from .metrics import (
+    calculate_iv_bonus,
+    calculate_raid_score,
+    iv_bonus,
+    raid_score,
+)
+
+__all__ = [
+    "calculate_iv_bonus",
+    "calculate_raid_score",
+    "iv_bonus",
+    "raid_score",
+]
